Make floodfill extend layerGroupPointer
